Update GitHub Actions dependencies
On one hand, this is a good idea in general. But also the code as it is written already relies on a newer interface than the version that was set before. This didn't come up as the current code isn't yet in use in the other actions, but would break the code if released like this.

Signed-off-by: Daniel Bartholomae <daniel.bartholomae@optilyz.com>
Signed-off-by: Daniel Bartholomae <daniel@bartholomae.name>