chore: use provided name to read previous report
Use custom artifact name when reading the previous report not always
using the old/hardcoded name otherwise it would miss the artifact and
lead to new issues always.
Also, upload the artifact with the custom name even if not raising
issues as following runs my start raising them.

Close #20.
See also #21, which is the same but before the refactoring to
TypeScript.

Signed-off-by: thc202 <thc202@gmail.com>